Personatester/requirements.txt
openai
pandas
tqdm
python-dotenv
